feat: move settings FeedFilters to commons
Additional filters moved to commons/ui/feeds:
- HiddenWordsFeedFilter
- HiddenAccountsFeedFilter (with ICacheProvider param)
- SpammerAccountsFeedFilter (with ICacheProvider param)

IAccount additions:
- liveHiddenUsers: LiveHiddenUsers

Part of KMP iOS migration tracked in #2238.
